ZON-6478: Move more IArea fields to z.contentquery.IConfiguration
