+++ core/src/zeit/contentquery/interfaces.py
@@ -5,6 +5,8 @@ import re
 import six
 import zc.sourcefactory.basic
 import zeit.cms.content.sources
+import zeit.content.cp.field
+import zeit.content.cp.source
 import zope.interface
 from zeit.cms.i18n import MessageFactory as _
 from zeit.cms.interfaces import CONFIG_CACHE
@@ -42,6 +44,67 @@ class AutomaticFeedSource(zeit.cms.content.sources.ObjectSource,
 
 
 AUTOMATIC_FEED_SOURCE = AutomaticFeedSource()
+
+
+class SimpleDictSource(zc.sourcefactory.basic.BasicSourceFactory):
+
+    values = collections.OrderedDict()
+
+    def getValues(self):
+        return self.values.keys()
+
+    def getTitle(self, value):
+        return self.values.get(value, value)
+
+
+class QuerySortOrderSource(SimpleDictSource):
+
+    values = collections.OrderedDict((
+        ('payload.workflow.date_last_published_semantic:desc',
+         _('query-sort-order-last-published-semantic')),
+        ('payload.document.last-semantic-change:desc',
+         _('query-sort-order-last-semantic-change')),
+        ('payload.document.date_first_released:desc',
+         _('query-sort-order-first-released')),
+    ))
+
+
+class QueryTypeSource(SimpleDictSource):
+
+    values = collections.OrderedDict([
+        ('channels', _('query-type-channels')),
+        ('serie', _('query-type-serie')),
+        ('product', _('query-type-product')),
+        ('ressort', _('query-type-ressort')),
+        ('genre', _('query-type-genre')),
+        ('authorships', _('query-type-authorships')),
+        ('access', _('query-type-access')),
+        ('content_type', _('query-type-content-type')),
+    ])
+
+
+class AutomaticTypeSource(SimpleDictSource):
+
+    values = collections.OrderedDict([
+        ('centerpage', _('automatic-area-type-centerpage')),
+        ('custom', _('automatic-area-type-custom')),
+        ('topicpage', _('automatic-area-type-topicpage')),
+        ('query', _('automatic-area-type-query')),
+        ('elasticsearch-query', _('automatic-area-type-elasticsearch-query')),
+        ('rss-feed', _('automatic-area-type-rss-feed'))
+    ])
+
+    def getToken(self, value):
+        # JS needs to use these values, don't MD5 them.
+        return value
+
+
+class QueryOperatorSource(SimpleDictSource):
+
+    values = collections.OrderedDict([
+        ('eq', _('query-operator-equal')),
+        ('neq', _('query-operator-notequal')),
+    ])
 
 
 class TopicpageFilterSource(zc.sourcefactory.basic.BasicSourceFactory,
@@ -88,6 +151,47 @@ class TopicpageFilterSource(zc.sourcefactory.basic.BasicSourceFactory,
         return value
 
 
+class QuerySubRessortSource(zeit.cms.content.sources.SubRessortSource):
+
+    def _get_parent_value(self, context):
+        # `context` is the IArea, which is adaptable to `parent_value_iface`
+        # ICommonMetadata, since it is adaptable to ICenterPage -- but of
+        # course we don't want to restrict the query subressort according to
+        # the CP's ressort. So we disable this validation here and rely on the
+        # fact that the widget will only offer matching subressorts anyway.
+        return None
+
+
+class IQueryConditions(zeit.content.article.interfaces.IArticle):
+
+    # ICommonMetadata uses a ReferenceField, which makes no sense for `query`.
+    authorships = zope.schema.Choice(
+        title=_("Authors"),
+        source=zeit.cms.content.interfaces.authorSource,
+        required=False)
+
+    # ICommonMetadata has ressort and sub_ressort in separate fields, but we
+    # need them combined. And so that whitespace-separated serializing works,
+    # we wrap it in a tuple to reuse the DAVPropertyConverter for `channels`.
+    ressort = zope.schema.Tuple(value_type=zc.form.field.Combination(
+        (zope.schema.Choice(
+            title=_('Ressort'),
+            source=zeit.cms.content.sources.RessortSource()),
+         zope.schema.Choice(
+             title=_('Sub ressort'),
+             source=QuerySubRessortSource(),
+             required=False)),
+        default=(),
+        required=False))
+    zope.interface.alsoProvides(
+        ressort.value_type, zeit.cms.content.interfaces.IChannelField)
+
+    # non-ICommonMetadata field
+    content_type = zope.schema.Choice(
+        title=_("Content type"),
+        source=zeit.cms.content.sources.CMSContentTypeSource())
+
+
 class IContentQuery(zope.interface.Interface):
     """Mechanism to retrieve content objects.
     Used to register named adapters for the different IArea.automatic_type's
@@ -109,6 +213,59 @@ class IContentQuery(zope.interface.Interface):
 
 
 class IConfiguration(zope.interface.Interface):
+    automatic_type = zope.schema.Choice(
+        title=_('automatic-area-type'),
+        source=AutomaticTypeSource(),
+        required=True)
+    automatic_type.__doc__ = """Determines from where IRenderedArea retrieves
+    content objects. Will look up a utility of that name for IContentQuery.
+    """
+
+    # XXX Rename to make clear that this setting only applies to AutoPilot.
+    count = zope.schema.Int(title=_('Amount of teasers'), default=15)
+
+    hide_dupes = zope.schema.Bool(
+        title=_('Hide duplicate teasers'),
+        default=True)
+
+    query = zope.schema.Tuple(
+        title=_('Custom Query'),
+        value_type=zeit.content.cp.field.DynamicCombination(
+            zope.schema.Choice(
+                title=_('Custom Query Type'),
+                source=QueryTypeSource(), default='channels'),
+            IQueryConditions,
+            zope.schema.Choice(
+                title=_('Custom Query Operator'),
+                source=QueryOperatorSource(), default='eq'),
+        ),
+        default=(),
+        required=False)
+
+    query_order = zope.schema.Choice(
+        title=_('Sort order'),
+        source=QuerySortOrderSource(),
+        default=u'payload.workflow.date_last_published_semantic:desc',
+        required=True)
+
+    elasticsearch_raw_query = zope.schema.Text(
+        title=_('Elasticsearch raw query'),
+        required=False)
+    elasticsearch_raw_order = zope.schema.TextLine(
+        title=_('Sort order'),
+        default=u'payload.document.date_first_released:desc',
+        required=False)
+    is_complete_query = zope.schema.Bool(
+        title=_('Take over complete query body'),
+        description=_('Remember to add payload.workflow.published:true'),
+        default=False,
+        required=False)
+
+    referenced_cp = zope.schema.Choice(
+        title=_('Get teasers from CenterPage'),
+        source=zeit.content.cp.source.centerPageSource,
+        required=False)
+
     referenced_topicpage = zope.schema.TextLine(
         title=_('Referenced Topicpage'),
         required=False)
